Fixed everything in R2
**Please don't delete this branch**
I may need to go back and change a few things later. For instance: 
Are we allowing a email to be used instead of strictly username when signing in?
Do I need Alt flow #.b in the use case? I don't want to overdo the use case again by talking about design (I.e.: "the session is saved in the database").