* Added TODO: game.makeMove() line in GamePage.java. This is what the line will switch to once it's all working.
***** It's all working, besides the interaction with the GUI!
Of course there are probably some bugs, but I wrote plenty of tests for plenty of interactions. Most if not all of the logic should be there.
The main issue is updating the board once we make a move. The game logic does it, but the GamePage does not update the drawing.
Need some help on this one from Marcel or Mike most likely.
* Game has a turn checker now. It will be deprecated in the future 100%, once players are gone. For now it works tho!
* There are some console outputs for sanity when making moves. You could play a game but youd have to use your imagination!